Move braille fktns
