+++ plotille/_util.py
@@ -25,8 +25,6 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 
 from math import floor
 
-import six
-
 
 def roundeven(x):
     '''Round to next even integer number in case of `X.5`
@@ -71,48 +69,3 @@ def hist(X, bins):  # noqa: N803
         y[x_idx] += 1
 
     return y, [i * xwidth + xmin for i in range(bins + 1)]
-
-
-def braille_from(dots):
-    '''Unicode character for braille with given dots set
-
-    See https://en.wikipedia.org/wiki/Braille_Patterns#Identifying.2C_naming_and_ordering
-    for dot to braille encoding.
-
-    Parameters:
-        dots: List[int]  All dots that should be set. Allowed dots are 1,2,3,4,5,6,7,8
-
-    Returns:
-        unicode: braille sign with given dots set. \u2800 - \u28ff
-    '''
-    bin_code = ['0'] * 8
-    for i in dots:
-        bin_code[8 - i] = '1'
-
-    code = 0x2800 + int(''.join(bin_code), 2)
-
-    return six.unichr(code)
-
-
-def dots_from(braille):
-    '''Get set dots from given
-
-    See https://en.wikipedia.org/wiki/Braille_Patterns#Identifying.2C_naming_and_ordering
-    for braille to dot decoding.
-
-    Parameters:
-        braille: unicode  Braille character in \u2800 - \u28ff
-
-    Returns:
-        List[int]: dots that are set in braille sign
-    '''
-    assert 0x2800 <= ord(braille) <= 0x28ff
-
-    code = six.text_type(bin(ord(braille) - 0x2800))[2:].rjust(8, '0')
-
-    dots = []
-    for i, c in enumerate(code):
-        if c == '1':
-            dots += [8 - i]
-
-    return sorted(dots)
